+++ packages/cloud_firestore/cloud_firestore/android/src/main/java/io/flutter/plugins/firebase/firestore/GeneratedAndroidFirebaseFirestore.java
@@ -2504,6 +2504,7 @@ public class GeneratedAndroidFirebaseFirestore {
     /** Failure case callback method for handling errors. */
     void error(@NonNull Throwable error);
   }
+
   /** Asynchronous error handling return type for nullable API method returns. */
   public interface NullableResult<T> {
     /** Success case callback method for handling returns. */
@@ -2512,6 +2513,7 @@ public class GeneratedAndroidFirebaseFirestore {
     /** Failure case callback method for handling errors. */
     void error(@NonNull Throwable error);
   }
+
   /** Asynchronous error handling return type for void API method returns. */
   public interface VoidResult {
     /** Success case callback method for handling returns. */
@@ -2520,6 +2522,7 @@ public class GeneratedAndroidFirebaseFirestore {
     /** Failure case callback method for handling errors. */
     void error(@NonNull Throwable error);
   }
+
   /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
   public interface FirebaseFirestoreHostApi {
 
@@ -2646,6 +2649,7 @@ public class GeneratedAndroidFirebaseFirestore {
     static @NonNull MessageCodec<Object> getCodec() {
       return PigeonCodec.INSTANCE;
     }
+
     /**
      * Sets up an instance of `FirebaseFirestoreHostApi` to handle messages through the
      * `binaryMessenger`.