+++ packages/cloud_firestore/cloud_firestore/ios/cloud_firestore/Package.swift
@@ -7,19 +7,19 @@
 
 import PackageDescription
 
-let library_version = "6.4.1"
-let firebase_sdk_version: Version = "12.13.0"
+let libraryVersion = "6.6.0"
+let firebaseSdkVersion: Version = "12.15.0"
 
 let package = Package(
   name: "cloud_firestore",
   platforms: [
-    .iOS("15.0"),
+    .iOS("15.0")
   ],
   products: [
-    .library(name: "cloud-firestore", targets: ["cloud_firestore"]),
+    .library(name: "cloud-firestore", targets: ["cloud_firestore"])
   ],
   dependencies: [
-    .package(url: "https://github.com/firebase/firebase-ios-sdk", from: firebase_sdk_version),
+    .package(url: "https://github.com/firebase/firebase-ios-sdk", exact: firebaseSdkVersion),
     .package(name: "firebase_core", path: "../firebase_core"),
   ],
   targets: [
@@ -30,14 +30,14 @@ let package = Package(
         .product(name: "firebase-core", package: "firebase_core"),
       ],
       resources: [
-        .process("Resources"),
+        .process("Resources")
       ],
       cSettings: [
         .headerSearchPath("include/cloud_firestore/Private"),
         .headerSearchPath("include/cloud_firestore/Public"),
-        .define("LIBRARY_VERSION", to: "\"\(library_version)\""),
+        .define("LIBRARY_VERSION", to: "\"\(libraryVersion)\""),
         .define("LIBRARY_NAME", to: "\"flutter-fire-fst\""),
       ]
-    ),
+    )
   ]
 )