+++ packages/cloud_firestore/cloud_firestore/ios/cloud_firestore/Sources/cloud_firestore/FLTPipelineParser.m
@@ -173,6 +173,16 @@ static NSError *parseError(NSString *message) {
     return [self parseExpression:(NSDictionary *)exprMap error:error];
   }
 
+  if ([name isEqualToString:@"document_matches"]) {
+    NSString *query = args[@"query"];
+    if (![query isKindOfClass:[NSString class]]) {
+      if (error) *error = parseError(@"document_matches requires query");
+      return nil;
+    }
+    FIRExprBridge *queryExpr = [[FIRConstantBridge alloc] init:query];
+    return FLTNewFunctionExprBridge(@"document_matches", @[ queryExpr ]);
+  }
+
   // Map Dart names to iOS SDK names where they differ
   NSString *sdkName = name;
   if ([name isEqualToString:@"bit_xor"]) sdkName = @"xor";
@@ -484,6 +494,86 @@ static NSError *parseError(NSString *message) {
       return nil;
     }
     return FLTNewFunctionExprBridge(@"array_concat", all);
+  }
+
+  // -------------------------------------------------------------------------
+  // expression + offset (+ optional length): array_slice
+  // -------------------------------------------------------------------------
+  if ([name isEqualToString:@"array_slice"]) {
+    id exprMap = args[@"expression"];
+    id offsetMap = args[@"offset"];
+    id lengthMap = args[@"length"];
+    if (![exprMap isKindOfClass:[NSDictionary class]] ||
+        ![offsetMap isKindOfClass:[NSDictionary class]]) {
+      if (error) *error = parseError(@"array_slice requires expression and offset");
+      return nil;
+    }
+    FIRExprBridge *expr = [self parseExpression:exprMap error:error];
+    FIRExprBridge *offset = [self parseExpression:offsetMap error:error];
+    if (!expr || !offset) return nil;
+    NSMutableArray<FIRExprBridge *> *sliceArgs =
+        [NSMutableArray arrayWithObjects:expr, offset, nil];
+    if ([lengthMap isKindOfClass:[NSDictionary class]]) {
+      FIRExprBridge *length = [self parseExpression:lengthMap error:error];
+      if (!length) return nil;
+      [sliceArgs addObject:length];
+    }
+    return FLTNewFunctionExprBridge(@"array_slice", sliceArgs);
+  }
+
+  // -------------------------------------------------------------------------
+  // expression + alias + filter: array_filter
+  // -------------------------------------------------------------------------
+  if ([name isEqualToString:@"array_filter"]) {
+    id exprMap = args[@"expression"];
+    NSString *alias = args[@"alias"];
+    id filterMap = args[@"filter"];
+    if (![exprMap isKindOfClass:[NSDictionary class]] || ![alias isKindOfClass:[NSString class]] ||
+        ![filterMap isKindOfClass:[NSDictionary class]]) {
+      if (error) *error = parseError(@"array_filter requires expression, alias, and filter");
+      return nil;
+    }
+    FIRExprBridge *expr = [self parseExpression:exprMap error:error];
+    FIRExprBridge *filter = [self parseBooleanExpression:filterMap error:error];
+    if (!expr || !filter) return nil;
+    return FLTNewFunctionExprBridge(@"array_filter",
+                                    @[ expr, [[FIRConstantBridge alloc] init:alias], filter ]);
+  }
+
+  // -------------------------------------------------------------------------
+  // expression + aliases + transform: array_transform / array_transform_with_index
+  // -------------------------------------------------------------------------
+  if ([name isEqualToString:@"array_transform"] ||
+      [name isEqualToString:@"array_transform_with_index"]) {
+    id exprMap = args[@"expression"];
+    NSString *elementAlias = args[@"element_alias"];
+    NSString *indexAlias = args[@"index_alias"];
+    id transformMap = args[@"transform"];
+    BOOL withIndex = [name isEqualToString:@"array_transform_with_index"];
+    if (![exprMap isKindOfClass:[NSDictionary class]] ||
+        ![elementAlias isKindOfClass:[NSString class]] ||
+        (withIndex && ![indexAlias isKindOfClass:[NSString class]]) ||
+        ![transformMap isKindOfClass:[NSDictionary class]]) {
+      if (error) {
+        NSString *message =
+            withIndex
+                ? @"array_transform_with_index requires expression, element_alias, index_alias, "
+                  @"and transform"
+                : @"array_transform requires expression, element_alias, and transform";
+        *error = parseError(message);
+      }
+      return nil;
+    }
+    FIRExprBridge *expr = [self parseExpression:exprMap error:error];
+    FIRExprBridge *transform = [self parseExpression:transformMap error:error];
+    if (!expr || !transform) return nil;
+    NSMutableArray<FIRExprBridge *> *transformArgs =
+        [NSMutableArray arrayWithObjects:expr, [[FIRConstantBridge alloc] init:elementAlias], nil];
+    if (withIndex) {
+      [transformArgs addObject:[[FIRConstantBridge alloc] init:indexAlias]];
+    }
+    [transformArgs addObject:transform];
+    return FLTNewFunctionExprBridge(name, transformArgs);
   }
 
   // -------------------------------------------------------------------------
@@ -828,6 +918,112 @@ static NSError *parseError(NSString *message) {
   return nil;
 }
 
++ (NSDictionary<NSString *, FIRExprBridge *> *)
+    parseSearchFieldsWithExpressionMaps:(NSArray<NSDictionary<NSString *, id> *> *)exprMaps
+                             exprParser:(FLTPipelineExpressionParser *)exprParser
+                                  error:(NSError **)error {
+  NSMutableDictionary<NSString *, FIRExprBridge *> *fields = [NSMutableDictionary dictionary];
+  NSError *parseErr = nil;
+
+  for (id em in exprMaps) {
+    if (![em isKindOfClass:[NSDictionary class]]) continue;
+
+    FIRExprBridge *expr = [exprParser parseExpression:em error:&parseErr];
+    if (!expr) {
+      if (error) *error = parseErr;
+      return nil;
+    }
+
+    NSString *key = [self keyForExpressionMap:em error:error];
+    if (![key isKindOfClass:[NSString class]] || key.length == 0) return nil;
+    fields[key] = expr;
+  }
+
+  return fields;
+}
+
++ (FIRStageBridge *)parseSearchStageWithArgs:(NSDictionary *)args
+                                  exprParser:(FLTPipelineExpressionParser *)exprParser
+                                       error:(NSError **)error {
+  NSString *queryType = args[@"query_type"];
+  id query = args[@"query"];
+  NSMutableDictionary<NSString *, FIRExprBridge *> *options = [NSMutableDictionary dictionary];
+  NSError *parseErr = nil;
+
+  if ([queryType isEqualToString:@"string"]) {
+    if (![query isKindOfClass:[NSString class]]) {
+      if (error) *error = parseError(@"search query_type 'string' requires string query");
+      return nil;
+    }
+    FIRExprBridge *queryExpr = [[FIRConstantBridge alloc] init:query];
+    options[@"query"] = FLTNewFunctionExprBridge(@"document_matches", @[ queryExpr ]);
+  } else if ([queryType isEqualToString:@"expression"]) {
+    if (![query isKindOfClass:[NSDictionary class]]) {
+      if (error) *error = parseError(@"search query_type 'expression' requires expression query");
+      return nil;
+    }
+    FIRExprBridge *queryExpr = [exprParser parseBooleanExpression:query error:&parseErr];
+    if (!queryExpr) {
+      if (error) *error = parseErr;
+      return nil;
+    }
+    options[@"query"] = queryExpr;
+  } else {
+    if (error) *error = parseError(@"search requires query_type to be 'string' or 'expression'");
+    return nil;
+  }
+
+  NSNumber *limit = [args[@"limit"] isKindOfClass:[NSNumber class]] ? args[@"limit"] : nil;
+  if (limit) options[@"limit"] = [[FIRConstantBridge alloc] init:limit];
+
+  NSNumber *offset = [args[@"offset"] isKindOfClass:[NSNumber class]] ? args[@"offset"] : nil;
+  if (offset) options[@"offset"] = [[FIRConstantBridge alloc] init:offset];
+
+  NSNumber *retrievalDepth =
+      [args[@"retrieval_depth"] isKindOfClass:[NSNumber class]] ? args[@"retrieval_depth"] : nil;
+  if (retrievalDepth) {
+    options[@"retrieval_depth"] = [[FIRConstantBridge alloc] init:retrievalDepth];
+  }
+
+  NSString *languageCode =
+      [args[@"language_code"] isKindOfClass:[NSString class]] ? args[@"language_code"] : nil;
+  if (languageCode) {
+    options[@"language_code"] = [[FIRConstantBridge alloc] init:languageCode];
+  }
+
+  NSMutableArray<FIROrderingBridge *> *sort = [NSMutableArray array];
+  NSArray *orderingMaps = args[@"sort"];
+  if ([orderingMaps isKindOfClass:[NSArray class]]) {
+    for (id om in orderingMaps) {
+      if (![om isKindOfClass:[NSDictionary class]]) continue;
+      id exprMap = ((NSDictionary *)om)[@"expression"];
+      NSString *dir = ((NSDictionary *)om)[@"order_direction"];
+      if (![exprMap isKindOfClass:[NSDictionary class]]) continue;
+      FIRExprBridge *expr = [exprParser parseExpression:exprMap error:&parseErr];
+      if (!expr) {
+        if (error) *error = parseErr;
+        return nil;
+      }
+      NSString *direction = [dir isEqualToString:@"asc"] ? @"ascending" : @"descending";
+      [sort addObject:[[FIROrderingBridge alloc] initWithExpr:expr Direction:direction]];
+    }
+  }
+
+  NSDictionary<NSString *, FIRExprBridge *> *addFields = @{};
+  NSArray *addFieldMaps = args[@"add_fields"];
+  if ([addFieldMaps isKindOfClass:[NSArray class]] && addFieldMaps.count > 0) {
+    addFields = [self parseSearchFieldsWithExpressionMaps:addFieldMaps
+                                               exprParser:exprParser
+                                                    error:error];
+    if (!addFields) return nil;
+  }
+
+  return [[FIRSearchStageBridge alloc] initWithOptions:options
+                                             addFields:addFields
+                                                select:@{}
+                                                  sort:sort];
+}
+
 + (NSArray<FIRStageBridge *> *)
     parseStagesWithFirestore:(FIRFirestore *)firestore
                       stages:(NSArray<NSDictionary<NSString *, id> *> *)stages
@@ -862,14 +1058,17 @@ static NSError *parseError(NSString *message) {
           return nil;
         }
         FIRCollectionReference *ref = [firestore collectionWithPath:path];
-        stage = [[FIRCollectionSourceStageBridge alloc] initWithRef:ref firestore:firestore];
+        stage = [[FIRCollectionSourceStageBridge alloc] initWithRef:ref
+                                                          firestore:firestore
+                                                         forceIndex:nil];
       } else if ([stageName isEqualToString:@"collection_group"]) {
         NSString *path = args[@"path"];
         if (!path) {
           if (error) *error = parseError(@"collection_group requires 'path'");
           return nil;
         }
-        stage = [[FIRCollectionGroupSourceStageBridge alloc] initWithCollectionId:path];
+        stage = [[FIRCollectionGroupSourceStageBridge alloc] initWithCollectionId:path
+                                                                       forceIndex:nil];
       } else if ([stageName isEqualToString:@"database"]) {
         stage = [[FIRDatabaseSourceStageBridge alloc] init];
       } else if ([stageName isEqualToString:@"documents"]) {
@@ -906,6 +1105,9 @@ static NSError *parseError(NSString *message) {
           return nil;
         }
         stage = [[FIRWhereStageBridge alloc] initWithExpr:expr];
+      } else if ([stageName isEqualToString:@"search"]) {
+        stage = [self parseSearchStageWithArgs:args exprParser:exprParser error:error];
+        if (!stage) return nil;
       } else if ([stageName isEqualToString:@"limit"]) {
         NSNumber *limit = args[@"limit"];
         if (![limit isKindOfClass:[NSNumber class]]) {