+++ packages/_flutterfire_internals/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.73
+
+ - Update a dependency to the latest release.
+
+## 1.3.72
+
+ - Update a dependency to the latest release.
+
 ## 1.3.71
 
  - Update a dependency to the latest release.