+++ packages/cloud_functions/cloud_functions_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.1.9
+
+ - Update a dependency to the latest release.
+
+## 5.1.8
+
+ - Update a dependency to the latest release.
+
 ## 5.1.7
 
  - Update a dependency to the latest release.