+++ packages/firebase_app_installations/firebase_app_installations_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.7+9
+
+ - Update a dependency to the latest release.
+
+## 0.1.7+8
+
+ - Update a dependency to the latest release.
+
 ## 0.1.7+7
 
  - Update a dependency to the latest release.