+++ packages/firebase_crashlytics/firebase_crashlytics_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.8.24
+
+ - Update a dependency to the latest release.
+
+## 3.8.23
+
+ - Update a dependency to the latest release.
+
 ## 3.8.22
 
  - Update a dependency to the latest release.