+++ packages/firebase_ml_model_downloader/firebase_ml_model_downloader_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.5+24
+
+ - Update a dependency to the latest release.
+
+## 0.1.5+23
+
+ - Update a dependency to the latest release.
+
 ## 0.1.5+22
 
  - Update a dependency to the latest release.