+++ packages/firebase_performance/firebase_performance_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.8+9
+
+ - Update a dependency to the latest release.
+
+## 0.1.8+8
+
+ - Update a dependency to the latest release.
+
 ## 0.1.8+7
 
  - Update a dependency to the latest release.