+++ packages/firebase_remote_config/firebase_remote_config_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.10.10
+
+ - Update a dependency to the latest release.
+
+## 1.10.9
+
+ - Update a dependency to the latest release.
+
 ## 1.10.8
 
  - Update a dependency to the latest release.