+++ packages/firebase_storage/firebase_storage/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 13.4.3
+
+ - Update a dependency to the latest release.
+
+## 13.4.2
+
+ - Update a dependency to the latest release.
+
 ## 13.4.1
 
  - Update a dependency to the latest release.