+++ packages/firebase_storage/firebase_storage_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.11.9
+
+ - Update a dependency to the latest release.
+
+## 3.11.8
+
+ - Update a dependency to the latest release.
+
 ## 3.11.7
 
  - Update a dependency to the latest release.